Move variadics to separate header.
